backend/requirements.txt
flask
flask-sqlalchemy
flask-cors
psycopg2-binary
pyjwt
python-dotenv
bcrypt
pandas
faker
pydantic
sqlalchemy
asyncpg
httpx
sqlglot
langgraph
langchain
chromadb
